Get commit id in interactive mode
When passing the commit as a single string (text) in the interactive
mode, the extractor now checks the beginning of that text (the first
line) for "commit" followed by a whitespace. If it finds this string, it
identifies the following characters until the next whitespace or a line
break as the id of that commit. If it does not find these elements, the
extraction process is terminated and an error message is printed.