Extend checks during commit id identification
If using the interactive mode, the respective extraction method tries to
identify the commit id in the first line of the given string. For the
highly unlikely case that this first line contains the expected
information, but is the only line passed to the extractor, an exception
was thrown that the id is not available, which is wrong. Now, the method
correctly identifies the id resulting in an empty commit (as passed by
the user in this case).