Fix Linux/OpenJDK-specific commit id bug
On Linux/OpenJDK-10.0.2 the individual commit numbers are surrounded by
additional quotation marks while reading the InputStream of the command
for retrieving the commit ids from the log. These additional marks
result in an error, if they are used in further commands, like for
retrieving the committer date, as git expects the commit id without any
surrounding characters. Hence, this fix simply replaces all occurrences
of quotation marks in the commit id list with an empty string before
this list is further processed.